// src/api/types/IApiClientApiDefaultsOptions.ts
export interface IApiClientApiDefaultsOptions {
  /**
   * Default model to use for completion requests
   */
  completionModel?: string

  /**
   * Default model to use for chat gpt3 8k tokens completion requests
   */
  chat3_8Model?: string

  /**
   * Default model to use for chat gpt3 16k tokens completion requests
   */
  chat3_16Model?: string

  /**
   * Default model to use for chat gpt4 8k tokens completion requests
   */
  chat4_8Model?: string

  /**
   * Default model to use for text edit requests
   */
  editTextModel?: string

  /**
   * Default model to use for code edit requests
   */
  editCodeModel?: string

  /**
   * If openai api returns more than one choice, this string will be used to delimit them.
   */
  choicesDelimiter?: string
}